+++ src/main/java/nl/jqno/equalsverifier/internal/prefabvalues/factoryproviders/JavaxFactoryProvider.java
@@ -3,6 +3,7 @@ package nl.jqno.equalsverifier.internal.prefabvalues.factoryproviders;
 import static nl.jqno.equalsverifier.internal.prefabvalues.factories.Factories.values;
 
 import javax.naming.Reference;
+import javax.swing.tree.DefaultMutableTreeNode;
 import nl.jqno.equalsverifier.internal.prefabvalues.FactoryCache;
 
 public final class JavaxFactoryProvider implements FactoryProvider {
@@ -13,6 +14,12 @@ public final class JavaxFactoryProvider implements FactoryProvider {
         cache.put(
                 Reference.class,
                 values(new Reference("one"), new Reference("two"), new Reference("one")));
+        cache.put(
+                DefaultMutableTreeNode.class,
+                values(
+                        new DefaultMutableTreeNode(),
+                        new DefaultMutableTreeNode(new Object()),
+                        new DefaultMutableTreeNode()));
 
         return cache;
     }